+++ de.bitub.step.p21/src/de/bitub/step/p21/P21IndexImpl.java
@@ -1,4 +1,4 @@
-package de.bitub.step.p21.util;
+package de.bitub.step.p21;
 
 import java.util.ArrayList;
 import java.util.Collection;
